DAP - Sub App - Asset User Community - QA - Bug Fix Android # 1 #4357
