+++ DAP/android/sub_app/fermat-dap-android-sub-app-asset-user-community-bitdubai/src/main/java/com/bitdubai/fermat_dap_android_sub_app_asset_user_community_bitdubai/factory/CommunityUserFragmentFactory.java
@@ -3,7 +3,7 @@ package com.bitdubai.fermat_dap_android_sub_app_asset_user_community_bitdubai.fa
 import com.bitdubai.fermat_android_api.engine.FermatFragmentFactory;
 import com.bitdubai.fermat_android_api.layer.definition.wallet.AbstractFermatFragment;
 import com.bitdubai.fermat_android_api.layer.definition.wallet.exceptions.FragmentNotFoundException;
-import com.bitdubai.fermat_dap_android_sub_app_asset_user_community_bitdubai.fragments.UserCmmuinityHomeFragment;
+import com.bitdubai.fermat_dap_android_sub_app_asset_user_community_bitdubai.fragments.UserCommuinityHomeFragment;
 import com.bitdubai.fermat_dap_android_sub_app_asset_user_community_bitdubai.fragments.UserCommuinityGroupUsersFragment;
 import com.bitdubai.fermat_dap_android_sub_app_asset_user_community_bitdubai.fragments.UserCommuinityUsersFragment;
 import com.bitdubai.fermat_dap_android_sub_app_asset_user_community_bitdubai.fragments.UserCommunityGroupFragment;
@@ -20,7 +20,7 @@ public class CommunityUserFragmentFactory extends FermatFragmentFactory<AssetUse
     public AbstractFermatFragment getFermatFragment(FragmentFactoryEnumType fragments) throws FragmentNotFoundException {
         switch (fragments) {
             case DAP_ASSET_USER_COMMUNITY_ACTIVITY_MAIN:
-                return UserCmmuinityHomeFragment.newInstance();
+                return UserCommuinityHomeFragment.newInstance();
             case DAP_ASSET_USER_COMMUNITY_ACTIVITY_ADMINISTRATIVE_GROUP_FRAGMENT:
                 return UserCommunityGroupFragment.newInstance();
             case DAP_ASSET_USER_COMMUNITY_ACTIVITY_ADMINISTRATIVE_GROUP_USERS_FRAGMENT:
