+++ DAP/android/sub_app/fermat-dap-android-sub-app-asset-user-community-bitdubai/src/main/java/com/bitdubai/fermat_dap_android_sub_app_asset_user_community_bitdubai/fragments/UserCommuinityGroupUsersFragment.java
@@ -53,7 +53,8 @@ import java.util.List;
 import static android.widget.Toast.makeText;
 
 /**
- * Home Fragment
+ * UserCommuinityGroupUsersFragment, Show all the users in the selected group
+ *
  */
 public class UserCommuinityGroupUsersFragment extends AbstractFermatFragment implements SwipeRefreshLayout.OnRefreshListener {
